move tests for AccountStateViewGasTracked
